rday.py: refactor hardcoded known_recycling_day variable for simplicity
initially wrote expecting future use to entail reading
the last known recycling date from a text file or ENV var
etc, rather than hard coding it into the script.
Just going for simplicity to get it working and if
reading the value from a string and then parsing that
into a date object is deemed desirable in future then
it can be added in future, but now is not the time to
be making provisions for future scenarios that might
never eventuate.